Learned Event handlers and how to bind them (see description)
for the button onClick, we are "not" calling the handleIncrement method but we're referencing it. When doing this we don't have access to it because it would need to be normally instantiated first like "obj A", but we want to avoid this. So we go back to what we learned previously and create a custom constructor and binding 'this' function so that it will return a new instance of itself when referenced.

Or even better; to just convert the said function to an arrow function which already inherits 'this' and we'll have access to the entire Counter object.