Learned the difference between States and Props, created a setter function to change a state of another component
props should only be read-only.
What you want to modify is a component's state, however, the state is only visible to the component and not seen by other components.

Rule of thumb: "The component that owns a piece of the state, should be the one modifying it."

Since the delete button was rendered in counter.jsx but the state we want to modify is in counters.jsx, we created