Here we've cleaned up how we change the state
when we had a lot of props to a class tag, it can become a cluster so we can just get the prop of a class entirely, so then we'd have all the state values of that properties without singlely naming each individual prop for each individual state value.

We've also did an argument pass in handleDelete() in counters.jsx to change the state by filtering out the counter.id that's been clicked.

The delete button works!