Made a reset button, it works only updates one source of data
there should be only one source where all data is stored.
How the counter is being incremented locally and how it's displayed in the browser has difference sources. So changing one source doesn't change the other. Next step would be to condense and reroute everything to one source of data.