Fixed an issue where there was multiple save state issues a mistype of an object syntax.
Here's I've also learned to have one source of data instead of multiple sources. Previously each component had their one state; one had the main array we were focusing on and the other state was using props.

So we deleted the state that was using props and switched all the button click calls to use this.props.functionName from the other component/class.

With that we were able to focus on only one state.

Then later I implemented the increment buttons by cloning the array, getting the index of which index was pressed, clone that index, increment that index, and then replace the main state by using setState().

Also a reminder note to self:
[...Array] is a spread operator to clone an array, which is similar to:
Array = new Array[size];