Lifted the state from counters.jsx to App.js, also transformed App function to Class App, And Navbar added to display total number of counter objects with value higher than 0
The goal here is to create a Navbar to display total number of active counters with their values higher than 0.

To do this I got a simple Navbar component from Bootstrap navbar examples and created a navbar.jsx component to render that navbar. Then I passed in the totalCounters that was created in App.js, and to get that I used "props".

In App.js the state(), handleReset(), handleIncrement(), and handleDelete(), from counters.jsx was copied over since we're moving the state() and its functions to a new parent(App.js).

App.js now displays the NavBar, has the functionality for <Counters>, and a new totalCounters variable to display in the NavBar that displays the length of the counters[] array if the value is higher than 0.