Learned how to do object destructuring
In navbar.jsx: Got rid of the Navbar class and instead used a NavBar const arrow function, the argument passed in was in {} curly braces to indicate that it's a destructure, meaning whatever is inside the {} already has props as a variable.

So passing in {totalCounters} is the same as this.props.totalCounters

I did the same in counters.jsx except we simply added a const destructure on the needed functions or wherever this.props was used in the code. So this gets rid of retyping 'this.props' to make the code cleaner.