Implemented the Decrement button with a check to see if the counter is 0, so that the decrement button gets greyed out
Rearranged the buttons in counter.jsx to align horizontally correctly using row, col, and m-2 to create margins.

I copied what I did with implement buttons and did the same to decrementing it but instead of adding it's subtracting, and to avoid going to the negative values;
this was added to the decrement button:

disabled ={this.props.counter.value === 0 ? "disabled" : ""}

meaning if the current counter value is 0 then disable is true, otherwise it's an empty string.